Add a quick benchmarking within cluster-features
This benchmark tests several variations of clustering, with distinct
number of clusters. For each k, the benchmark presents the inertia of
the clustering, and the silhouette score. The first value indicates how
well the area of the clustering is divided, and the second how "precise"
is the clustering. Less variation in inertia indicates a better
clustering. Similarly, silhouette values near 1 also indicates the
quality of the clustering.

For a complete description of these metrics, please refer to the
SciKit-Learn documentation.